Fixed empowered stage to be a byte.
Added the ability to pass empowered stage into spell cast.
Added ISpellCalculateBonusCoefficient to calc a spells base damage before any calculations.

+++ Source/Game/DataStorage/Structs/S_Records.cs
@@ -336,7 +336,7 @@ namespace Game.DataStorage
     public sealed class SpellEmpowerStageRecord
     {
         public uint Id;
-        public uint Stage;
+        public byte Stage;
         public uint DurationMs;
         public uint SpellEmpowerID;
     }
