+++ Source/Game/Spells/SpellInfo.cs
@@ -4048,7 +4048,7 @@ namespace Game.Spells
         public SpellSchoolMask SchoolMask { get; set; }
         public uint ChargeCategoryId;
         public List<uint> Labels = new();
-        public Dictionary<uint, SpellEmpowerStageRecord> EmpowerStages { get; set; } = new();
+        public Dictionary<byte, SpellEmpowerStageRecord> EmpowerStages { get; set; } = new();
 
         // SpellScalingEntry
         public ScalingInfo Scaling;
